security: replace unsafe-inline CSP with per-request nonces
Generate a UUID-based nonce per request in SecurityHeadersService,
pass it through request extensions to page handlers, and inject it
into Tera template contexts. All inline <script> and <style> tags
now require the matching nonce attribute, eliminating the need for
'unsafe-inline' in the Content-Security-Policy header.

Also fix a pre-existing statvfs type mismatch in backup-manager on
platforms where f_bavail/f_frsize have different widths.

+++ plugins/backup-manager/src/space.rs
@@ -68,7 +68,7 @@ pub fn available_space(path: &Path) -> Result<u64> {
         if libc::statvfs(c_path.as_ptr(), stat.as_mut_ptr()) == 0 {
             let stat = stat.assume_init();
             // Available space for unprivileged users
-            Ok(stat.f_bavail.saturating_mul(stat.f_frsize))
+            Ok(u64::from(stat.f_bavail).saturating_mul(stat.f_frsize))
         } else {
             Err(VoomError::Io(std::io::Error::last_os_error()))
         }
